Refactor search file into utils and pull_requests_search
